+++ src/contrib/transient-ref.function.mts
@@ -1,5 +1,5 @@
 import { TYPE_WRAPPER, TypeWrapperStamp } from '../constants/type-wrapper.const.mjs';
-import { TransientRefHost } from '../transient-ref-host.class';
+import { TransientRefHost } from './transient-ref-host.class.mjs';
 import { InferServiceType } from '../types/infer-service-type.type.mjs';
 import { ServiceIdentifier } from '../types/service-identifier.type.mjs';
 import { TypeWrapper } from '../types/type-wrapper.type.mjs';
